+++ client-device-connection-infinispan/src/test/java/org/eclipse/hono/deviceconnection/infinispan/client/QuarkusPropertyBindingTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -16,9 +16,14 @@ package org.eclipse.hono.deviceconnection.infinispan.client;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import javax.inject.Inject;
+import java.util.List;
 
-import org.infinispan.client.hotrod.impl.ConfigurationProperties;
+import javax.inject.Inject;
+import javax.security.sasl.Sasl;
+
+import org.infinispan.client.hotrod.configuration.ClusterConfiguration;
+import org.infinispan.client.hotrod.configuration.Configuration;
+import org.infinispan.client.hotrod.configuration.ConfigurationBuilder;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.junit.QuarkusTest;
@@ -52,19 +57,40 @@ public class QuarkusPropertyBindingTest {
     void testRemoteCacheConfigurationPropertiesArePickedUp() {
         assertThat(remoteCacheOptions).isNotNull();
         final var remoteCacheConfig = new InfinispanRemoteConfigurationProperties(remoteCacheOptions);
+        final Configuration builtRemoteCacheConfig = new ConfigurationBuilder()
+                .withProperties(remoteCacheConfig.getProperties()).build(false);
+
         assertThat(remoteCacheConfig.getServerList()).contains("data-grid:11222");
         assertThat(remoteCacheConfig.getAuthUsername()).isEqualTo("user");
         assertThat(remoteCacheConfig.getAuthPassword()).isEqualTo("secret");
         assertThat(remoteCacheConfig.getAuthRealm()).isEqualTo("ApplicationRealm");
+
+        final List<ClusterConfiguration> clusters = builtRemoteCacheConfig.clusters();
+        assertThat(clusters.size()).isEqualTo(2);
+        assertThat(clusters.get(0).getClusterName()).isEqualTo("siteA");
+        assertThat(clusters.get(0).getCluster().size()).isEqualTo(2);
+        assertThat(clusters.get(0).getCluster().get(0).host()).isEqualTo("hostA1");
+        assertThat(clusters.get(0).getCluster().get(0).port()).isEqualTo(11222);
+        assertThat(clusters.get(1).getClusterName()).isEqualTo("siteB");
+
+        assertThat(remoteCacheConfig.gtConnectionPoolMinIdle()).isEqualTo(10);
+        assertThat(remoteCacheConfig.getConnectionPoolMaxActive()).isEqualTo(10);
+        assertThat(remoteCacheConfig.getConnectionPoolMaxPendingRequests()).isEqualTo(400);
+        assertThat(remoteCacheConfig.getConnectionPoolMaxWait()).isEqualTo(500);
+
+        assertThat(remoteCacheConfig.getDefaultExecutorFactoryPoolSize()).isEqualTo(200);
+
         assertThat(remoteCacheConfig.getSaslMechanism()).contains("DIGEST-MD5");
+        final var saslProperties = builtRemoteCacheConfig.security().authentication().saslProperties();
+        assertThat(saslProperties.get(Sasl.QOP)).isEqualTo("auth");
+
         assertThat(remoteCacheConfig.getSoTimeout()).isEqualTo(5000);
         assertThat(remoteCacheConfig.getConnectTimeout()).isEqualTo(5000);
         assertThat(remoteCacheConfig.getKeyStoreFileName()).isEqualTo("/etc/hono/key-store.p12");
         assertThat(remoteCacheConfig.getKeyStoreType()).isEqualTo("PKCS12");
         assertThat(remoteCacheConfig.getKeyStorePassword()).isEqualTo("key-store-secret");
         assertThat(remoteCacheConfig.getKeyAlias()).isEqualTo("infinispan");
-        assertThat(remoteCacheConfig.getProperties().getProperty(ConfigurationProperties.KEY_STORE_CERTIFICATE_PASSWORD))
-            .isEqualTo("cert-secret");
+        assertThat(remoteCacheConfig.getKeyStoreCertificatePassword()).isEqualTo("cert-secret");
         assertThat(remoteCacheConfig.getTrustStoreFileName()).isEqualTo("/etc/hono/trust-store-file.p12");
         assertThat(remoteCacheConfig.getTrustStorePath()).isEqualTo("/etc/hono/trust-store.p12");
         assertThat(remoteCacheConfig.getTrustStoreType()).isEqualTo("PKCS12");